refactor(dashboard,api): split webview window and run event stream
